requirements.txt
aiogram==3.13.1
motor==3.6.0
pymongo==4.10.1
httpx==0.27.2
python-dotenv==1.0.1
apscheduler==3.10.4
aiohttp==3.10.11
